+++ count/count_traces.h
